third_party/blink/web_tests/platform/linux/virtual/webnn-service-with-gpu/external/wpt/webnn/conformance_tests/maxPool2d.https.any_gpu-expected.txt

This is a testharness.js-based test.
[FAIL] [required] maxPool2d float32 4D tensor options.dilations
  promise_test: Unhandled rejection with value: object "NotSupportedError: Failed to execute 'build' on 'MLGraphBuilder': Pool2d in tflite doesn't support dilations."
[FAIL] [required] maxPool2d float32 4D tensor options.outputShapeRounding=ceil and symmetric padding
  assert_less_than_equal: assert_array_approx_equals_ulp: test maxPool2d float32 actual -10000 should be close enough to expected 0 by ULP distance: expected a number less than or equal to 0n but got 1176256512n
[FAIL] [required] maxPool2d float16 4D tensor options.dilations
  promise_test: Unhandled rejection with value: object "NotSupportedError: Failed to execute 'build' on 'MLGraphBuilder': Pool2d in tflite doesn't support dilations."
Harness: the test ran to completion.
